Fix narrowing-conversion warnings surfaced by ubsan; unblock tsan build
The ubsan/tsan sanitizer presets are Debug builds whose instrumentation
defeats GCC's value-range propagation, exposing latent implicit narrowings
that -Werror=conversion/-Werror=sign-conversion then make fatal:

- byte-pair little-endian reads `a | (b << 8)` integer-promote to int and
  narrow on assignment to uint16_t (zlib_anongame.cpp, protocol d2cs/d2dbs
  fsm.cpp, two tests) -- wrapped each in an explicit static_cast<uint16_t>.
- migrate_accounts.cpp hex-nibble accumulators were `unsigned int` assigned
  from int subtractions (sign-conversion) -> use int; the final byte is
  already explicitly cast.

These narrowings are value-preserving by construction; the casts just make
that explicit so the warning is satisfied in all builds.

Also unblock the tsan build: GCC's -Wtsan fires on libstdc++'s own
std::atomic_thread_fence (inside shared_ptr's atomic refcount), which TSan
cannot model -- a standard-library tooling limitation, not our code. Demote it
to a warning via -Wno-error=tsan in pvpgn_v3_target_werror (inert outside tsan).

ubsan now builds clean; affected tests pass (285/285 in d2cs/d2dbs/anongame/
compression/migrate).

+++ src/infra/compression/src/zlib_anongame.cpp
@@ -77,10 +77,16 @@ core::Result<std::vector<std::uint8_t>> anongame_decompress(
             core::StatusCode::OutOfRange,
             "anongame_decompress: input shorter than 4-byte header"));
     }
-    const std::uint16_t raw_len = static_cast<std::uint16_t>(framed[0]) |
-                                  (static_cast<std::uint16_t>(framed[1]) << 8);
-    const std::uint16_t comp_len = static_cast<std::uint16_t>(framed[2]) |
-                                   (static_cast<std::uint16_t>(framed[3]) << 8);
+    // The shift integer-promotes each operand to int, so the OR is an int that
+    // narrows on assignment to uint16_t. Both bytes provably fit, but make the
+    // narrowing explicit — value-range propagation suppresses the -Wconversion
+    // warning in plain builds, yet ubsan instrumentation defeats that analysis.
+    const std::uint16_t raw_len = static_cast<std::uint16_t>(
+        static_cast<std::uint16_t>(framed[0]) |
+        (static_cast<std::uint16_t>(framed[1]) << 8));
+    const std::uint16_t comp_len = static_cast<std::uint16_t>(
+        static_cast<std::uint16_t>(framed[2]) |
+        (static_cast<std::uint16_t>(framed[3]) << 8));
     if (framed.size() - kAnonGameHeaderSize < comp_len) {
         return core::fail(core::make_error(
             core::StatusCode::OutOfRange,
